// src/app/model/status_consulta.ts
export interface IStatus {
  valor: string;
}

export enum StatusConsulta {
  AGUARDANDO_TRIAGEM = 'AGUARDANDO_TRIAGEM',
  AGUARDANDO_ATENDIMENTO = 'AGUARDANDO_ATENDIMENTO',
  NECESSITA_MEDICACAO = 'NECESSITA_MEDICACAO',
  RETORNO_MEDICO = 'RETORNO_MEDICO',
  ATENDIMENTO_FINALIZADO = 'ATENDIMENTO_FINALIZADO'
}